🌱 split unified/types: organize into subdirectory
