+++ web/src/lib/unified/types/dashboard.types.ts
@@ -1,0 +1,2 @@
+// Dashboard types can be extended here if needed
+export type { UnifiedCardConfig } from './card.types'
