+++ web/src/lib/unified/types/index.ts
@@ -1,0 +1,4 @@
+// Re-export all unified types from split files for organized imports
+export * from './card.types'
+export type { } from './statblock.types'
+export type { } from './dashboard.types'
