+++ web/src/lib/unified/types/statblock.types.ts
@@ -1,0 +1,2 @@
+// StatBlock types are defined in card.types.ts but re-exported here for organized imports
+export type { CardStatConfig, CardStatAction } from './card.types'
